pyproject.toml
[project]
name = "ros2-gateway-test"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "matplotlib>=3.10.5",
    "mcap-ros2-support>=0.5.5",
    "numpy>=2.3.2",
    "pandas>=2.3.1",
    "seaborn>=0.13.2",
]
